# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Example Swagger setup for Golang Gin.",
        "title": "Assistant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9876",
    "basePath": "/",
    "paths": {
...
        "/wiki/entries": {
            "get": {
                "description": "分页获取知识库条目列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "获取知识库条目列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "返回数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的知识库条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "创建知识库条目",
                "parameters": [
                    {
                        "description": "条目信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_modules_wiki.createEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/wiki/entries/{id}": {
            "get": {
                "description": "根据ID获取知识库条目详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "获取知识库条目详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "条目不存在",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除知识库条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "删除知识库条目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/wiki/import": {
            "post": {
                "description": "从Markdown文件导入知识库条目",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "导入知识库条目",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Markdown文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/wiki/search": {
            "get": {
                "description": "根据关键词搜索知识库条目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库"
                ],
                "summary": "搜索知识库条目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "返回数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "assistant_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_modules_wiki.createEntryRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "keywords": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}